wasm/jsapi/esm-integration/counter-from-wasm.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type=module>
setup({ single_test: true });
import {count, increment} from "./counter-from-wasm.wasm";
assert_equals(count.value, 5);
increment();
assert_equals(count.value, 6);
done();
</script>
